A StateMachine
Simple state machine that uses an integer as its state and processes
message synchronously.

I have a simple test which uses cpuTime to test passing a new message
and allocating a new message each time. Allocating takes about 3000 times
longer at about 5us per message where as passing the same message takes
1.7ns.

The takeaway here is that allocating a new message is expensive, and
that we need to resue them if possible.