Add messagearena
This is not thread safe nor shareable between threads
but my simple tests seem to work. As I say in the
messagearena.nim header I don't like the casting especially
sometimes I'm casting the same value in one instance as pointer
and others not as a pointer. There's got to be a better way!