Added tsq and use Ptr's and no Ref's
tsq is a thread safe queue initial implementation not fully tested

When developing tsq I found that when a Ptr is casted to a Ref there
is an automatic reference counter incremented and a runtime error
occurs. I surprised it didn't happen before, but it seems just using
Ptr's is the way to go.