Default to two MsgLoopers.
Sometimes t5 hangs when using one MsgLooper so change to use two but
allow to test with one. Using one most of the time works but I've seen
it not work. The solution is probably to use createThread as noted in
this nim form thread: http://forum.nim-lang.org/t/959#5825