Use createThread
The bug mentioned in the previous looks to be that I wasn't using both
message loopers and when I did it broke hard because there was I had
a recursive lock attempt.

The solution here uses two loops and processes the messages in one loop
where looper does not hold any locks. Then if no messages were processed
in the first loop then a lock is held during the second loop and a new
proc, msgqueue.emptyNolock, was used.

This isn't very efficient and there are certainly better algorithms,
for instance implement a recursive lock would be one.

This code also uses a local dbg proc to make debugging easier although
the technique didn't work for TSm.processMsg. I do see there is a
"module logging" in nim, I'll have to see how that works.