Use --passL:-lttng-ust to instead of a separate link execution of gcc.
For some reason without this change I was getting a "log" statement
undefined when compiling with PTP (public tracepoint) true. I'd recently
seen a suggestion on the forum that "passL" could be used to get other
libraries. This seemed to do the trick so using that.

Also, removed --listCmd and added to nim.cfg, although commented out.